Refactor PclusterConfig
Now we have a config folder containing 4 modules:
* mapping.py
* param_types.py
* pcluster_config.py: corresponds to the old cfnconfig.py
* validators.py: corresponds to the old config_sanity.py

It contains a map of all the sections and the parameters configurable by the user in the configuration file.

For each section you can define:
* type, the class to use to represent this section (default: Section)
*key, the key used in configuration file that identifies the section type
(e.g [cluster default] -> "cluster" is the key)
*label, the label used in configuration file that identifies a specific section
(e.g [cluster default] -> "default" is the key)
*validator, a function to use to validate the section.
It is called for all the parameters once all of them are initialized.
*cfn, the CFN parameters to use for the to/from_cfn conversion
(it is used for sections that are converted to a single CFN parameter, e.g. RAID, EFS, FSX)
*params, a dictionary containing all the parameters available for that section

For each parameter you can define:
* type the class to use to represent this section (default: Param, a string parameter)
* cfn the CFN parameters to use for the to/from_cfn conversion.
* allowed_values, a list of allowed values or a regex
* validator, a function to use to validate the param.
It is called for all the parameters once all of them are initialized.
* default, a default value for the internal representation,
if not specified the value will be None
* referred_section, it is a special attribute used only for *SettingsParam,
the parameters that refers to other sections in the file (e.g. vpc_settings)

It contains the classes to manage all the parameters and the sections,
the standard ones (e.g. Param, IntParam, FloatParam, etc) and
the custom ones (e.g. SharedDirParam, SpotPriceParam, etc).

I defined a new custom *Param and *Section Class if the "to/from_file" or "to/from_cfn" conversion
require custom actions (e.g. merge/check values from multiple configuration parameters).

It contains the class to manage a pcluster configuration.
It only contains a single cluster configuration and it can be initialized
from map (default values), from file or from CFN parameters.

It contains all the validation function.

I created 800+ tests, to verify *Section, *Param, PclusterConfig classes and the validators.
I also added a test_source_consistency.py to verify if the values used in the tests
are the default ones used in the CloudFormation template and to check the number of parameters.

To add new section you have to double check the PclusterConfig initialization.
If you add a new parameters that will change the default values
for the internal representation or if it is converted to a CFN input parameter,
you'll have to update the defaults.py in the tests folder.

* Added underscore prefix to all the private functions in commands.py (old pcluster.py)
* Added get_stack_name function to add prefix to cluster-name for both awsbatch and pcluster CLIs
* cli: Don't use cluster template from config file for start/stop commands
* cli: Improved tags, extra json and extra-parameters help
* don't pass region as input of the PclusterConfig initialization and set region and AWS credentials in the environment so we don't need to pass to all the boto3 calls
* _check_for_updates function has been moved outside the PclusterConfig class, into the create command, since it is called only at creation time.
* template_url is not longer an input of the PclusterConfig class, and it has been moved into the create command.
* tags are not longer an input of the PclusterConfig and are merged to the ones from the configuration file at creation time
* sanity check done by validators is now able to distinguish errors from warnings

Signed-off-by: Enrico Usai <usai@amazon.com>