Create ebs section only when SharedDir CFN param contains commas
We can initialize a single EBS volume by using the shared_dir config
parameter only, or by using a single [ebs] section with all the default
values except the shared_dir one contained in the ebs section.

The NumberOfEBSVol CFN param is 1 in both cases.
When parsing CFN to create the configuration we can differentiate the
two cases by verifying if the SharedDir parameter contains commas or not.

+ Added tests to verify the two cases

Signed-off-by: Enrico Usai <usai@amazon.com>