Add NetworkCardIndex parameter to Launch Templates
The new parameter `NetworkCardIndex` is required for instances types with more than one Network Card; it allows to specify to which card a NetworkInterface is assigned
The default value for this parameter is '0' which must be assigned to the primary Network Interface. For this reason we specify this property only to secondary interfaces, if used; this solution allows to keep supporting instance types with single network cards with the new templates until the support for this new parameter in CloudFormation will be generally available.

Signed-off-by: ddeidda <ddeidda@amazon.com>